moved test data storage event method to new file in dataservice package
